CombatComp TryAttack 호출
Fire에서 Combat Comp의 Try Attack을 호출 해 볼 수 있었다

+++ TimeThiefServer/Content/Gameplay/Combat/CombatTypes.h
@@ -24,7 +24,7 @@ struct AttackRequest
     SE::Math::Vector3 direction{};
    
     float range = 0.0f;
-    float damage = 0.0f;
+    int32 damage = 0;   // 공격이 가할 피해량 (향후 무기 시스템이 구현되면 활용)
     
     uint32 weaponId = 0;   // 공격에 사용된 무기의 ID (향후 무기 시스템이 구현되면 활용)
 };
